requirements.txt
playwright
pytest
allure-pytest
requests
pytest-asyncio
asyncio
```